+++ src/pools/cluster/fixed.ts
@@ -73,10 +73,7 @@ export class FixedClusterPool<
     worker.on('disconnect', () => {
       worker.kill()
     })
-    await this.sendKillMessageToWorker(
-      workerNodeKey,
-      workerNode.info.id as number
-    )
+    await this.sendKillMessageToWorker(workerNodeKey)
     worker.disconnect()
     await waitWorkerExit
   }
@@ -86,14 +83,16 @@ export class FixedClusterPool<
     workerNodeKey: number,
     message: MessageValue<Data>
   ): void {
-    this.workerNodes[workerNodeKey].worker.send(message)
+    this.workerNodes[workerNodeKey].worker.send({
+      ...message,
+      workerId: this.workerNodes[workerNodeKey].info.id as number
+    })
   }
 
   /** @inheritDoc */
   protected sendStartupMessageToWorker (workerNodeKey: number): void {
     this.sendToWorker(workerNodeKey, {
-      ready: false,
-      workerId: this.workerNodes[workerNodeKey].info.id as number
+      ready: false
     })
   }